+++ src/index.ts
@@ -1,0 +1,2 @@
+export { fetchChannelMessages, fetchGuildMessages } from './main.js';
+export { Fetcher } from './Fetcher';
