+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
 	"compilerOptions": {
 		"alwaysStrict": true,
+		"allowJs": true,
 		"declaration": true,
 		"declarationDir": "types",
 		"lib": ["ESNext"],
@@ -9,10 +10,10 @@
 		"noImplicitAny": true,
 		"outDir": "dist",
 		"pretty": true,
-		"rootDir": "src",
 		"skipLibCheck": true,
 		"strict": true,
 		"target": "esNext"
 	},
-	"exclude": ["node_modules"]
+	"exclude": ["node_modules"],
+	"files": ["src/main.js", "src/index.ts", "src/Fetcher.ts"]
 }
